# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /publications:
    get:
      operationId: PublicationsController_list
      summary: Список публикаций
      description: "Публикации по задачам, где пользователь owner или executor.
        Создаются автоматически при `COMPLETED` задачи. Фильтры: `role`,
        `postId`, `taskId`, `q` (title)."
      parameters:
        - name: role
          required: false
          in: query
          description: owner — публикации по моим задачам; executor — где я исполнитель
          schema:
            enum:
              - owner
              - executor
            type: string
        - name: postId
          required: false
          in: query
          schema:
            format: uuid
            type: string
        - name: taskId
          required: false
          in: query
          schema:
            format: uuid
            type: string
        - name: q
          required: false
          in: query
          description: Поиск по title
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
      responses:
        "200":
          description: Список публикаций с пагинацией
      tags: &a18
        - publications
      security: &a19
        - access-token: []
  /publications/{id}:
    get:
      operationId: PublicationsController_findById
      summary: Публикация по id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationResponseDto"
        "403":
          description: Нет доступа
        "404":
          description: Публикация не найдена
      tags: *a18
      security: *a19
    patch:
      operationId: PublicationsController_update
      summary: Обновить публикацию
      description: Участники задачи могут редактировать `title`, `description`,
        `externalUrl`, `platform`.
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePublicationDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationResponseDto"
        "403":
          description: Нет доступа
        "404":
          description: Публикация не найдена
      tags: *a18
      security: *a19
...
info:
  title: Project Back API
  description: >-
    # Project Back API


    Backend для marketplace creator ↔ company. Документация для фронтенда и
    AI-ассистентов.


    ## Модель данных


    | Сущность | Что это |

    |----------|---------|

    | **Account** | Login: email + password. Один человек = один Account. |

    | **User** | Публичный профиль: CREATOR или COMPANY. При регистрации
    создаётся один User на Account. |

    | **Membership** | Доступ Account к User: OWNER / ADMIN / EDITOR / VIEWER. |

    | **Invite** | Приглашение другому человеку управлять **существующим**
    профилем. |


    **Активный профиль** — тот, чей `userId` в JWT (`sub`). Переключение: `POST
    /auth/switch-profile`.


    ## Cookies и CORS


    - После login/register/refresh выставляются httpOnly cookies:
    `access-token`, `refresh-token`.

    - Все запросы с `credentials: 'include'`.

    - CORS origin: `CORS_ORIGIN` (обычно http://localhost:3000).


    ## Сценарии


    ### Первый вход

    1. `POST /auth/register/creator` или `/auth/register/company` — создаёт
    Account + первый профиль.

    2. `POST /auth/login` — вход по email/password.


    ### Совместное управление (менеджер)

    1. Владелец: `POST /auth/invites` — email менеджера + `userId`
    **существующего** профиля + роль.

    2. Менеджер: регистрация/логин **своим** email → `POST
    /auth/invites/accept`.

    3. Менеджер: `GET /auth/profiles` → `POST /auth/switch-profile`.


...

...

...


...

...

...


...

...

...

...

...
servers:
...
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: access-token
  schemas:
...
    ApplicationCreatorProfileDto:
      type: object
      properties:
        name:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
      required:
        - name
        - lastName
    ApplicationCompanyProfileDto:
      type: object
      properties:
        companyName:
          type: string
          example: Acme Inc.
      required:
        - companyName
    ApplicationOwnerDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        avatar:
          type: string
          example: https://example.com/avatar.jpg
        creatorProfile:
          $ref: "#/components/schemas/ApplicationCreatorProfileDto"
        companyProfile:
          $ref: "#/components/schemas/ApplicationCompanyProfileDto"
      required:
        - id
...
    ApplicationApplicantDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - CREATOR
            - COMPANY
        avatar:
          type: string
          nullable: true
        name:
          type: string
          example: John
          description: Для CREATOR
        lastName:
          type: string
          example: Doe
          description: Для CREATOR
        companyName:
          type: string
          example: Acme Inc.
          description: Для COMPANY
      required:
        - id
        - role
...
    PublicationMediaDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sourceTaskMediaId:
          type: string
          format: uuid
          nullable: true
        url:
          type: string
        key:
          type: string
        size:
          type: string
        mimeType:
          type: string
      required:
        - id
        - url
        - key
        - size
        - mimeType
    PublicationResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        postId:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        description:
          type: string
        externalUrl:
          type: string
          nullable: true
        platform:
          type: string
          enum:
            - INSTAGRAM
            - TIKTOK
            - YOUTUBE
            - TELEGRAM
            - VK
            - OTHER
          nullable: true
        brief:
          type: object
          additionalProperties: true
          nullable: true
        deliverables:
          type: object
          additionalProperties: true
          nullable: true
        location:
          type: object
          additionalProperties: true
          nullable: true
        status:
          type: string
          enum:
            - PUBLISHED
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        media:
          type: array
          items:
            $ref: "#/components/schemas/PublicationMediaDto"
        owner:
          $ref: "#/components/schemas/ApplicationOwnerDto"
        executor:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ApplicationApplicantDto"
      required:
        - id
        - taskId
        - postId
        - description
        - status
        - publishedAt
        - createdAt
        - updatedAt
        - media
        - owner
    UpdatePublicationDto:
      type: object
      properties:
        title:
          type: string
          nullable: true
        description:
          type: string
          maxLength: 5000
        externalUrl:
          type: string
          nullable: true
          example: https://instagram.com/p/abc
        platform:
          type: string
          enum:
            - INSTAGRAM
            - TIKTOK
            - YOUTUBE
            - TELEGRAM
            - VK
            - OTHER
          nullable: true
...